Evaluate token in session controller

+++ src/controllers/session-controller.php
@@ -4,19 +4,22 @@
  */ 
 class SessionController extends ControllerBase
 {
-  private function createHash($password)
-  {
-    return hash('md5', $password);
-  }
-
   // Get all sessions from db
   // URL: /api/session/active
   public function active()
   {
     // Create query finding all active sessions
-    $query = $this->entityManager->createQuery('SELECT s.id, s.name, s.isPrivate, count(m.id) memberCount  FROM Session s LEFT JOIN s.members m WHERE s.lastAction > ?1 GROUP BY s.id');
+    $query = $this->entityManager->createQuery('SELECT s.id, s.name, s.isPrivate, s.token, count(m.id) memberCount  FROM Session s LEFT JOIN s.members m WHERE s.lastAction > ?1 GROUP BY s.id');
     $query->setParameter(1, new DateTime('-1 hour'));
     $sessions = $query->getArrayResult();
+
+    // Determine password requirement for each session
+    foreach($sessions as &$session) {
+      $tokenKey = $this->tokenKey($session["id"]);
+      $session["requiresPassword"] = $session["isPrivate"] 
+        && (!isset($_COOKIE[$tokenKey]) || $_COOKIE[$tokenKey] !== $session["token"]);
+    }
+
     return $sessions;
   }
   
@@ -30,21 +33,20 @@ class SessionController extends ControllerBase
     $session->setName($data["name"]);
     $session->setCardSet($data["cardSet"]);
 
+    // Generate the access token and assign it to the session
     $private = $data["isPrivate"];
     $session->setIsPrivate($private);
     if ($private)
-      $session->setPassword($this->createHash($data["password"]));
-
-    // Generate the access token and assign it to the session
-    $token = bin2hex(random_bytes(16));
+      $token = $this->createHash($data["name"], $data["password"]);
+    else
+      $token = $this->createHash($data["name"], bin2hex(random_bytes(8)));   
     $session->setToken($token);
       
     $session->setLastAction(new DateTime());
 
     $this->save($session);
     
-    $tokenKey = $this->tokenKey($session->getId());
-    setcookie($tokenKey, $token, time()+60*60*24*30);
+    $this->setCookie($session);
     
     return new NumericResponse($session->getId());
   }
@@ -55,9 +57,8 @@ class SessionController extends ControllerBase
   {
     $method = $_SERVER['REQUEST_METHOD'];
     if ($method == "PUT")
-    { 
-      $data = $this->jsonInput();        
-      return $this->addMember($sessionId, $data["name"]);
+    {           
+      return $this->addMember($sessionId);
     }
     if ($method == "DELETE")
     {
@@ -66,10 +67,15 @@ class SessionController extends ControllerBase
   }
   
   // Add a member with this name to the session
-  private function addMember($id, $name)
+  private function addMember($id)
   {   
-    $session = $this->getSession($id);
+    $data = $this->jsonInput();
+    $name = $data["name"];
     
+    $session = $this->getSession($id);
+    $token = $session->getToken();
+    $tokenKey = $this->tokenKey($session->getId());
+
     // Check for existing member
     foreach($session->getMembers() as $candidate)
     {
@@ -78,6 +84,27 @@ class SessionController extends ControllerBase
         $member = $candidate;
         break;
       }  
+    }
+
+    // This blocks check different ways to be granted access
+    if ($this->tokenProvided($session, $name)) {
+      // The user already has the token so we do nothing and continue
+    } else if(isset($_GET["token"]) && $_GET["token"] == $token) {
+      // User supplied the token
+      $this->setCookie($session);
+    } else if(isset($data["password"]) && $token === $this->createHash($session->getName(), $data["password"])) {
+      // Or the password
+      $this->setCookie($session);
+    } else if ($session->getIsPrivate() == false && !isset($member)) {
+      // Tokens are only required to join private sessions,
+      // but without the token the rights are restricted by a member-only token
+      // unless this member already exists
+      $memberToken = $this->createHash($name, $token);
+      $this->setCookie($session, $memberToken);
+    } else {
+      // Return access forbidden otherwise
+      http_response_code(403); // Return HTTP 403 FORBIDDEN
+      return;
     }
 
     // Create new member
@@ -90,11 +117,11 @@ class SessionController extends ControllerBase
         
       $this->saveAll([$member, $session]);
     }
+
+    // Set name cookie for faster login next time
+    setcookie('scrum_member_name', $name, time()+60*60*24*30, "/");
     
-    // Store name in cookie if not done yet
-    // if(!isset($_COOKIE['scrum_member_name']))
-    //   setcookie('scrum_member_name', $name); European privacy law
-    
+    // Create response
     $result = new stdClass();
     $result->sessionId = $id;
     $result->memberId = $member->getId();
@@ -104,15 +131,19 @@ class SessionController extends ControllerBase
   // Remove member from session
   private function removeMember($id)
   {
-    include __DIR__ .  "/session-evaluation.php";
+    // Get member and session
+    $member = $this->getMember($id); 
+    $session = $member->getSession();
 
-    // Get and remove member
-    $member = $this->getMember($id);    
+    if (!$this->verifyToken($session, $member->getName()))
+      return;
+
+    // Get and remove member       
     $this->entityManager->remove($member);
     $this->entityManager->flush();
 
-    // Get the members session and its current poll
-    $session = $member->getSession();
+    // Reevaluate the current poll
+    include __DIR__ .  "/session-evaluation.php";
     $poll = $session->getCurrentPoll();
     if($poll !== null && SessionEvaluation::evaluatePoll($session, $poll))
     {
@@ -128,11 +159,14 @@ class SessionController extends ControllerBase
   }
   
   // Check if session is protected by password
-  // URL: /api/session/haspassword/{id}
-  public function haspassword($id)
+  // This only returns true of the requesting user does not
+  // have the token cookie
+  // URL: /api/session/requiresPassword/{id}
+  public function requiresPassword($id = 0)
   {
     $session = $this->getSession($id);
-    return new BoolResponse($session->getIsPrivate());
+    $requires = $session->getIsPrivate() && !$this->tokenProvided($session);
+    return new BoolResponse($requires);
   }
 
   // Check if member is still part of the session
@@ -154,7 +188,12 @@ class SessionController extends ControllerBase
   {
     $data = $this->jsonInput();
     $session = $this->getSession($id);
-    $result = $session->getPassword() === $this->createHash($data["password"]);
+    $result = $session->getToken() === $this->createHash($session->getName(), $data["password"]);
+
+    // If the correct password was transmitted we grant the token as a reward
+    if ($result)
+      $this->setCookie($session);
+
     return new BoolResponse($result);
   }
 
@@ -171,6 +210,18 @@ class SessionController extends ControllerBase
   public function cardsets()
   {
     return $this->cardSets;
+  }
+
+  // Set the token cookie for this session 
+  // with additional parameters for expiration and path
+  private function setCookie($session, $token = null)
+  {
+    $tokenKey = $this->tokenKey($session->getId());
+
+    if ($token == null)
+      $token = $session->getToken();
+
+    setcookie($tokenKey, $token, time()+60*60*24*30, "/");
   }
 }
